EA IFF-85 docs as published by Commodore
These are the original IFF docs, as published by Commodore Business
Machines. Thankfully they converted the docs to MS Word format which
still works today. The docs do contain some revisions from Commodore.

(The original docs are in MacWrite format which is not easy to open
today.)

I've converted them to Word *.docx format and from there to *.rtf and
*.pdf for longevity.